Move utils tests to own package
